refactor(AuthenticatedApp): unstack AuthenticatedApp router
